+++ src/UriHelper.php
@@ -4,6 +4,10 @@ namespace Rokka\Client;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
+use Rokka\Client\Core\Stack;
+use Rokka\Client\Core\StackOperation;
+use Rokka\Client\Core\StackUri;
+use Rokka\Client\Core\UriComponents;
 
 class UriHelper
 {
@@ -13,8 +17,8 @@ class UriHelper
      * Useful eg. if you just want to add "options-dpr-2" to an existing URL.
      * Returns the original URL, if it can't parse it as valid Rokka URL.
      *
-     * @param string $url     The rokka image render URL
-     * @param string $options The options you want to add as string
+     * @param string       $url     The rokka image render URL
+     * @param array|string $options The options you want to add as string
      *
      * @return string
      */
@@ -30,75 +34,218 @@ class UriHelper
      * Returns the original URL, if it can't parse it as valid Rokka URL.
      *
      * @param UriInterface $uri     The rokka image render URL
-     * @param string       $options The options you want to add as string
+     * @param array|string $options The options you want to add as string
      *
      * @return UriInterface
      */
     public static function addOptionsToUri(UriInterface $uri, $options)
     {
-        $path = $uri->getPath();
-        if (preg_match('#^/(?<stack>[^/]+)/(?<rest>[0-9a-f]{6,40}.*)$#', $path, $matches)) {
-            // nothing to do here
-        } elseif (preg_match('#^/(?<stack>[^/]+)/(?<options>[^/]+)/(?<rest>[0-9a-f]{6,40}.*)$#', $path, $matches)) {
-            $urlOptions = self::decomposeOptions($matches['options']);
-            $inputOptions = self::decomposeOptions($options);
-            $combinedOptions = array_replace_recursive($urlOptions, $inputOptions);
-            $newOptions = [];
-            foreach ($combinedOptions as $key => $values) {
-                $newOption = "$key";
-                foreach ($values as $k => $v) {
-                    $newOption .= "-$k-$v";
-                }
-                $newOptions[] = $newOption;
-            }
-            $options = implode('--', $newOptions);
-        } else {
+        if (is_array($options)) {
+            return self::addOptionsToUri($uri, self::getUriStringFromStackConfig($options));
+        }
+        $matches = self::decomposeUri($uri);
+        if (empty($matches)) {
             //if nothing matches, it's not a proper rokka URL, just return the original uri
             return $uri;
         }
+        $stack = $matches->getStack();
+        $stack->addOverridingOptions($options);
 
-        return $uri->withPath('/'.$matches['stack'].'/'.$options.'/'.$matches['rest']);
+        return self::composeUri($matches, $uri);
     }
 
     /**
-     * @param string $options
+     * Generate a rokka uri with an array or an UriComponent returned by decomposeUri().
      *
-     * @return array
+     * The array looks like
+     * ```
+     * ['stack' => 'stackname', #or StackUri object
+     *  'hash' => 'hash',
+     *  'filename' => 'filename-for-url',
+     *  'format' => 'image format', # eg. jpg
+     * ]
+     * ```
+     *
+     * @since 1.2.0
+     *
+     * @param array|UriComponents $components
+     * @param UriInterface        $uri        If this is provided, it will change the path for that object and return
+     *
+     * @return UriInterface
      */
-    private static function decomposeOptions($options)
+    public static function composeUri($components, UriInterface $uri = null)
     {
-        $components = [];
-        foreach (explode('--', $options) as $stringOperation) {
-            $stringOperationWithOptions = explode('-', $stringOperation);
-            $stringOperationName = $stringOperationWithOptions[0];
-            if ('' == $stringOperationName) {
-                continue;
+        if (is_array($components)) {
+            $components = UriComponents::createFromArray($components);
+        }
+        $stack = $components->getStack();
+        $stackName = $stack->getName();
+        $path = '/'.$stackName;
+        $stackConfig = $stack->getConfigAsArray();
+        $stackUrl = self::getUriStringFromStackConfig($stackConfig);
+        if (!empty($stackUrl)) {
+            $path .= '/'.$stackUrl;
+        }
+        if (!empty($components->getHash())) {
+            $path .= '/'.$components->getHash();
+
+            if (!empty($components->getFilename())) {
+                $path .= '/'.$components->getFilename();
             }
-            $components[$stringOperationName] = self::parseOptions(array_slice($stringOperationWithOptions, 1));
+
+            $path .= '.'.$components->getFormat();
         }
 
-        return $components;
+        if (null !== $uri) {
+            return  $uri->withPath($path);
+        }
+
+        return new Uri($path);
     }
 
     /**
-     * @param array $options
+     * Return components of a rokka URL.
      *
-     * @return array
+     * @since 1.2.0
+     *
+     * @param UriInterface $uri
+     *
+     * @return UriComponents|null
      */
-    private static function parseOptions(array $options)
+    public static function decomposeUri(UriInterface $uri)
     {
-        $optionValues = array_filter($options, function ($key) {
-            return $key % 2;
-        }, ARRAY_FILTER_USE_KEY);
+        $stackPattern = '(?<stack>.*([^-]|--)|-*)';
+        $hashPattern = '(?<hash>[0-9a-f]{6,40})';
+        $filenamePattern = '(?<filename>[A-Za-z\-\0-\9]+)';
+        $formatPattern = '(?<format>.{3,4})';
+        $pathPattern = '(?<hash>-.+-)';
 
-        $optionKeys = array_filter($options, function ($key) {
-            return !($key % 2);
-        }, ARRAY_FILTER_USE_KEY);
+        // hash with seo-filename
+        if (preg_match('#^/'.$stackPattern.'/'.$hashPattern.'/'.$filenamePattern.'\.'.$formatPattern.'$#', $uri->getPath(), $matches) ||
+            // hash without seo-filename
+            preg_match('#^/'.$stackPattern.'/'.$hashPattern.'.'.$formatPattern.'$#', $uri->getPath(), $matches) ||
+            // remote_path with seo-filename
+            preg_match('#^/'.$stackPattern.'/'.$pathPattern.'/'.$filenamePattern.'\.'.$formatPattern.'$#', $uri->getPath(), $matches) ||
+            // remote_path without seo-filename
+            preg_match('#^/'.$stackPattern.'/'.$pathPattern.'.'.$formatPattern.'$#', $uri->getPath(), $matches)) {
+            return UriComponents::createFromArray($matches);
+        }
+    }
 
-        if (count($optionKeys) !== count($optionValues)) {
-            throw new \InvalidArgumentException('The options given has to be an even array with key and value.');
+    /**
+     * @param string      $url    The original rokka render URL to be adjusted
+     * @param string      $size   The size of the image, eg '300w' or '2x'
+     * @param null|string $custom Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     *
+     * @return UriInterface
+     */
+    public static function getSrcSetUrlString($url, $size, $custom = null)
+    {
+        return self::getSrcSetUrl(new Uri($url), $size, $custom);
+    }
+
+    /**
+     * Returns a rokka URL to be used in srcset style attributes.
+     *
+     * $size can be eg. "2x" or "500w"
+     * $custom can be any rokka options you want to optionally add, or also a dpi identifier like "2x"
+     *
+     * This method will then generate the right rokka URLs to get what you want, see
+     * `\Rokka\Client\Tests\UriHelperTest::provideGetSrcSetUrl` for some examples and the expected returns.
+     *
+     * @param UriInterface $url    The original rokka render URL to be adjusted
+     * @param string       $size   The size of the image, eg '300w' or '2x'
+     * @param null|string  $custom Any rokka options you'd like to add, or are a dpi identifier like '2x'
+     *
+     * @return UriInterface
+     */
+    public static function getSrcSetUrl(UriInterface $url, $size, $custom = null)
+    {
+        $identifier = substr($size, -1, 1);
+        $size = substr($size, 0, -1);
+        if ('x' === $identifier) {
+            $uri = self::addOptionsToUri($url, 'options-dpr-'.$size);
+        } elseif ('w' === $identifier) {
+            $uri = self::addOptionsToUri($url, 'resize-width-'.$size);
+        } else {
+            return $url;
+        }
+        if (null !== $custom) {
+            if (preg_match('#^([0-9]+)x$#', $custom, $matches)) {
+                $uri = self::addOptionsToUri($uri, 'options-dpr-'.$matches[1].'--resize-width-'.(int) ceil($size / $matches[1]));
+            } else {
+                $stack = new StackUri();
+                $stack->addOverridingOptions($custom);
+                // if dpr is given in custom option, but not width, calculate correct width
+                $resizeOperations = $stack->getStackOperationsByName('resize');
+                $widthIsNotSet = true;
+                foreach ($resizeOperations as $resizeOperation) {
+                    if (isset($resizeOperation->options['width'])) {
+                        $widthIsNotSet = false;
+                    }
+                }
+                $options = $stack->getStackOptions();
+                if (isset($options['dpr']) && $widthIsNotSet) {
+                    $custom .= '--resize-width-'.(int) ceil($size / $options['dpr']);
+                }
+
+                $uri = self::addOptionsToUri($uri, $custom);
+            }
         }
 
-        return array_combine($optionKeys, $optionValues);
+        return $uri;
+    }
+
+    /**
+     * @param array $config
+     *
+     * @return string
+     */
+    private static function getUriStringFromStackConfig(array $config)
+    {
+        $newOptions = [];
+        if (isset($config['operations'])) {
+            foreach ($config['operations'] as $values) {
+                if ($values instanceof StackOperation) {
+                    $newOptions[] = self::getStringForOptions($values->name, $values->options);
+                } else {
+                    $newOptions[] = self::getStringForOptions($values['name'], $values['options']);
+                }
+            }
+        }
+
+        $newStackOptions = null;
+        if (isset($config['options'])) {
+            $newStackOptions = self::getStringForOptions('options', $config['options']);
+        }
+        //don't return this, if it's only "options" as string
+        if ('options' !== $newStackOptions) {
+            $newOptions[] = $newStackOptions;
+        }
+        $options = implode('--', $newOptions);
+
+        return $options;
+    }
+
+    /**
+     * @param string $name
+     * @param array  $values
+     *
+     * @return string
+     */
+    private static function getStringForOptions($name, $values)
+    {
+        $newOption = $name;
+        ksort($values);
+        foreach ($values as $k => $v) {
+            if (false === $v) {
+                $v = 'false';
+            } elseif (true === $v) {
+                $v = 'true';
+            }
+            $newOption .= "-$k-$v";
+        }
+
+        return $newOption;
     }
 }